new option button in the application page (#76)
