+++ tailwind.config.js
@@ -42,6 +42,7 @@ module.exports = {
       },
     },
   },
+  
   plugins: [],
 };
